+++ lib/capybara/chromedriver/logger.rb
@@ -12,31 +12,6 @@ module Capybara
     module Logger
       extend self
 
-      def build_capabilities(loggingPrefs: { browser: 'ALL' }, **options)
-        options[:chromeOptions] ||= {}
-
-        if options[:chromeOptions][:w3c]
-          warn "warning: Setting chromeOptions.w3c to true makes it not "\
-            "possible to get console.log messages from Chrome.\n\n"\
-            "Please see: https://github.com/SeleniumHQ/selenium/issues/7270"
-        else
-          options[:chromeOptions][:w3c] = false
-        end
-
-        if loggingPrefs[:browser] != 'ALL'
-          warn "warning: loggingPrefs needs to contain { browser: 'ALL' } "\
-            "when using Logger#build_capabilities"
-        end
-
-        options[:loggingPrefs] = loggingPrefs
-
-        # Support Chrome 75+
-        # see: https://github.com/SeleniumHQ/selenium/issues/7342
-        options["goog:loggingPrefs"] = loggingPrefs
-
-        ::Selenium::WebDriver::Remote::Capabilities.chrome(options)
-      end
-
       def filters
         @filters || []
       end
@@ -50,7 +25,7 @@ module Capybara
       end
 
       def filter_levels=(filters)
-        @filter_levels = filters && filters.map(&:upcase).map(&:to_s)
+        @filter_levels = filters&.map(&:upcase)&.map(&:to_s)
       end
 
       def raise_js_errors?