+++ spec/capybara/chromedriver/logger_spec.rb
@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 RSpec.describe Capybara::Chromedriver::Logger do
-  it "has a version number" do
+  it 'has a version number' do
     expect(Capybara::Chromedriver::Logger::VERSION).not_to be nil
   end
 end